apimas-drf/requirements.txt
Django>=1.8,<1.9
djangorestframework>=3.3.3
django-filter>=0.12
apimas
